chore: move parseError function
